+++ config/README.md
@@ -22,9 +22,9 @@ go get -u github.com/go-kratos/kratos/registry/nacos/v2
 
 ## kube
 ```shell
-go get -u github.com/go-kratos/kratos/registry/kube/v2
+go get -u github.com/go-kratos/kratos/registry/kubernetes/v2
 ```
 
 ```shell
-go get -u github.com/go-kratos/kratos/config/kube/v2
+go get -u github.com/go-kratos/kratos/config/kubernetes/v2
 ```
