+++ registry/kubernetes/go.mod
@@ -1,4 +1,4 @@
-module github.com/go-kratos/kratos/registry/kube/v2
+module github.com/go-kratos/kratos/registry/kubernetes/v2
 
 go 1.17
